+++ cloud-application/src/main/java/it/evsmartcharging/DatabaseModels.java
@@ -4,7 +4,7 @@ import java.sql.Timestamp;
 
 public class DatabaseModels {
 
-    // Represents a row from the plate_priority table
+    // Represents a row from the platePriority table
     public static class PlatePriority {
         private final String plate;
         private final int priority;
@@ -31,9 +31,9 @@ public class DatabaseModels {
     // Represents a row from the realPower table
     public static class RealPower {
         private final Timestamp timestamp;
-        private final int pReal;
+        private final Float pReal;
 
-        public RealPower(Timestamp timestamp, int pReal) {
+        public RealPower(Timestamp timestamp, Float pReal) {
             this.timestamp = timestamp;
             this.pReal = pReal;
         }
@@ -42,7 +42,7 @@ public class DatabaseModels {
             return timestamp;
         }
 
-        public int getPReal() {
+        public Float getPReal() {
             return pReal;
         }
 
@@ -55,14 +55,14 @@ public class DatabaseModels {
     // Represents a row from the solarData table
     public static class SolarData {
         private final Timestamp timestamp;
-        private final int Gb;
-        private final int Gd;
-        private final int Gr;
-        private final int HSun;
-        private final int T;
-        private final int WS;
+        private final Float Gb;
+        private final Float Gd;
+        private final Float Gr;
+        private final Float HSun;
+        private final Float T;
+        private final Float WS;
 
-        public SolarData(Timestamp timestamp, int Gb, int Gd, int Gr, int HSun, int T, int WS) {
+        public SolarData(Timestamp timestamp, Float Gb, Float Gd, Float Gr, Float HSun, Float T, Float WS) {
             this.timestamp = timestamp;
             this.Gb = Gb;
             this.Gd = Gd;
@@ -73,12 +73,12 @@ public class DatabaseModels {
         }
 
         public Timestamp getTimestamp() { return timestamp; }
-        public int getGb() { return Gb; }
-        public int getGd() { return Gd; }
-        public int getGr() { return Gr; }
-        public int getHSun() { return HSun; }
-        public int getT() { return T; }
-        public int getWS() { return WS; }
+        public Float getGb() { return Gb; }
+        public Float getGd() { return Gd; }
+        public Float getGr() { return Gr; }
+        public Float getHSun() { return HSun; }
+        public Float getT() { return T; }
+        public Float getWS() { return WS; }
 
         @Override
         public String toString() {
